change LCProps file to CommonProps
