--- packages/frontend-logic/src/types/LayoutComponentProps.ts
@@ -1,61 +1,0 @@
-import { UiSettingDT } from "@m-cafe-app/utils";
-import {
-  MouseEventHandler,
-  ChangeEventHandler,
-  FocusEventHandler
-} from "react";
-
-export interface CommonLCProps {
-  classNameOverride?: string;
-  classNameAddon?: string;
-  id?: string;
-  uiSettings?: {
-    [key: string]: UiSettingDT[]
-  };
-  label?: string;
-  tooltip?: string;
-}
-
-/**
- * Input fields props start
- */
-
-interface CommonFieldLCProps extends CommonLCProps {
-  placeholder?: string;
-  name: string;
-  value: string;
-  onChange: ChangeEventHandler<HTMLInputElement>;
-  onBlur: FocusEventHandler<HTMLInputElement>;
-  errorMessage?: string;
-  disabled?: boolean;
-}
-
-export interface TextFieldLCProps extends CommonFieldLCProps {
-  type?: string;
-}
-
-export interface TimeFieldLCProps extends CommonFieldLCProps {
-  svgUrl: string;
-  type?: string;
-}
-
-export interface TextAreaFieldLCProps extends CommonFieldLCProps {
-  maxrows: number
-}
-
-export interface SelectFieldLCProps extends CommonFieldLCProps {
-  svgUrl: string;
-  type?: string;
-  options: string[];
-  tNode?: string;
-  onChooseOption: MouseEventHandler<HTMLDivElement>;
-}
-
-export interface DateFieldLCProps extends CommonFieldLCProps {
-  svgUrl: string;
-  type?: string;
-}
-
-/**
- * Input fields props end
- */
